refactor(Grammar): let set detect invalid symbols
If Grammar is passed an invalid symbol, Python's set will detect
it and raise a TypeError. We don't need an extra layer of code
to detect and translate these errors into a different exception.

Note, we would need this extra layer of code if this was a user
error that we were trying to detect and report to the user. But
this is an internal programmer error that should be detected
and reported by unit tests.