refactor(Grammar): any iterable is a valid form
The second parameter to addRule() must be an iterable. With this change,
a str passed as a second parameter would result in each character being
treated as a symbol. If this was not the intent of the programmer,
then this is an error. However, this error should be detected by unit
tests.