refactor: consolidate shared data structures (#78)
We have many functions that operate on common data structures. The
definition of these data structures are next to the functions that
create them. For example, the Line structure is next to the function
that parses a line and creates an instance of Line.

The challenge is that when you need to reference the implementation of a
structure, you have to find where it first created in the system. It
would be better if there was a single place to look for shared
structures.

This commit creates a structs.py and errors.y in plcc.load_spec, and
moves all data structures and errors into them. This changes the imports
of MANY files.

There will likely need to be a couple of tuneups after this as we learn
more about the design.