refactor: remove ValidationError3 (#80)
When I moved all errors to plcc/load_spec/errors.py I found 3
implementations of ValidationError. To allow the refactoring tool to
work and make the move, I first renamed two of the implementations to
ValidationError2 and ValidationError3. Once the move was made, I
compared the implementations and found that the original and 2 were the
same, so I renamed ValidationError2 (and all its references) to just
ValidationError. Then I removed it leaving the original.

ValidationError3's implementation is different. It's constructor takes a
different parameter. This commit removes ValidationError3 and has each
of its subclass call the superclass constructor with the normal expected
parameters. No changes are needed by client code except their import
statements had to be adjusted to use ValidationError instead of
ValidationError3.

Closes #79