refactor: remove unused default parameters (#81)
In the beginning, all functions that used a regex accepted a default
parameter as the regex. I thought it would be nice to allow these to be
passed by the caller in case the syntax ever changed. For example...

```python
def parseTerminal(pattern=r'[A-Z_]+'):
    ...
```

This overcomplicates the function, and makes you wonder if there are any
call sites that are passing a different regex.

These should be rewriting to something like the following:

```python
def parseTerminal():
    pattern = r'[A-Z_]+'
    ...
```

Closes #75