refactor: clear out includes in __init__.py (#35)
I could not detect a clear pattern or practice of
imports in __init__.py, and they do not look well
maintained. So I've decided to delete all imports
in __init__.py

This may lead to some ugly imports. I suggest these
should be a "smell" that indicate that a refactor
may be warranted.