refactor(lexvalidate): each validator process all rules
Some validators are stateful across rules
(e.g., check_for_duplicate_names). Some are not.
If validators are assumed to process a single rule,
then to allow for stateful validators, validators
must be objects. This imposes a more complex protocol
for using validators.

Instead, if we move the responsibility of processing
all rules to each validator, then validators can
be represented by a simple function that takes an
iterable of rules, and returns a list of errors.

So that's what this commit does. Now all lexvalidate
validators are functions that each process all the
rules and return a list of errors.