refactor(spec.rough): add handler param to resolve_includes (#45)
Currently rough raises exceptions when errors found.
Instead we want to hand back a list of errors during
processing, and to try to continue processing as much
as possible.

But rough is implemented as a composable set of generator
returning functions. It doesn't make sense for it to
return a list of errors along with a generator, because
the error list would always be initially empty;
and the error list would fill as the generator is
expanded. This would be confusing and would likely
lead to complicated error prone code in clients.

Instead, we'll add an optional callback parameter
to each parse function. This callback will be called
and passed an error object each time an error is detected.
Assuming the callback doesn't raise an exception,
processing will continue after the callback returns.
This will allow clients to choose how to handle errors.

This commit is the first step along this path.
It adds a `handler` callback parameter to `resolve_includes`.
For example, a client can collect errors during processing
but allow processing to continue as follows.

```python
errors = []
generator = resolve_includes(..., handler=lambda e: errors.append(e))
````

If a handler is not provided, resolve_includes raises the first
error it encounters (the current behavior).