refactor(spec.rough): add handler param to parse_blocks (#46)
The default handler raises an exception (existing behavior).

If a handler is passed, and no closing %%% is found,
then parse_blocks produces a Block containing all of the
lines from the opening to the end of the file, plus an
additional line that parse_blocks creates to represent
the closing of the block.