refactor(spec.rough): add handler param to parse_from_*
This completes our refactor to add handler param
to spec.rough. Callers can now pass a `handler` callback
to parse_from_* functions. The handler is called
for each error encountered during the processing.
If the handler raises and exception, then processing
is halted and the exception is raised. Otherwise
the parser does its best to continue.

```python
from plccng.spec.rough import parse_from_string
results = list(parse_from_string(..., handler=lambda e: print(e)))
```

There are currently two possible errors when parsing
the rough. CircularIncludeError and UnclosedBlockError.
When a CircularIncludeError is encountered, the handler
is called, the Include is skipped, and processing continues.
When an UnclosedBlockError is encountered, the handler
is called, a closing is created to finish the Block,
and processing continues.

In the future, when new errors are added, the creator
must decide how processing can continue when that
error is detected.